crypto: explicitly initialize openssl
Previously, openssl was an optional dependency, still it
required initialization when it was loaded. This was solved by using
an __attribute__((constructor)) function executed when libsyslog-ng-crypto.so
was loaded.

As now openssl is a hard dependency and libsyslog-ng-crypto.so was merged
back into the main shared object, we don't need this trickery.

Also, with this change it might be possible not to initialize openssl
for example in unit tests that will get rid off a number of nasty
one-off initializations that can create a cleaner valgrind output.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>