cfg-parser: remove version specific keyword support
This patch removes version specific keyword support as it causes difficult
to resolve problems in SCL macros. The only motivation behind
this was to avoid conflicts between user-specified identifiers (e.g. source,
destination names and similar stuff) and keywords.

This is resolved by making the core keyword set as minimal as possible and
if there's still a conflict, we can always create one-off fixes to address
them individually.

The original problem was related to the "null" keyword used by SLES
installations by default, the null keyword is now delegated to afsql and
csvparser as they are the only one using it. No need to keep it in the
core set.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>