configure.ac: install tools and scl under a syslog-ng specific subdirectory
These should never be installed in /usr/share directly, but rather under a
subdirectory and as described in

https://www.gnu.org/prep/standards/html_node/Directory-Variables.html

we should do that right within the source and not rely on packaging tools to
do it for us.

This will trigger a required change in packaging scripts to avoid changing
the --datadir, as the default of /usr/share should work out-of-the-box.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>