Signed-off-by: Tibor Benke <tibor.benke@balabit.com>
java: build the language binding only with make

This commit allows us to build the language bindings without Gradle.

The following steps are involved:
1. generate class files
2. generate header files based on class files
3. add class files to syslog-ng-core.jar

The class files are built with javac. Some input source files has class
files as dependencies (e.g. LogPipe.java), some are leaf files (e.g.
InternalMessageSender.java). I checked the dependencies between the files
are created the corresponding *_CLASS_DEPS variables. There is no need to
define the dependencies (undefined variables in make expand to the empty
string), if a source file is a leaf (without any dependency).

The .SECONDEXPANSION is needed to use these dependency lists based on the
stem (the matching part). So if make builds LogPipe.class, the $* expands
to `LogPipe`. The $ symbol has to be prefixed with an other $ in
second expansions.

The .gradle directory don't have to be removed, as Gradle won't use
this directory.

It's worth to check the manual pages of javac and javah.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>