Disk-buffer: this kind of disk buffer is ported from syslog-ng PE.
This is an inner-destination plugin, make it possible to use a disk based queue.
There are 2 kind of disk-buffer: reliable and non-reliable.

Non-reliable means, that it uses memory buffers,
therefore this kind of disk-buffer is not crash safe,
but in some cases it can be faster than the reliable one

The reliable disk-buffer store every message on the queue in serialized format,
therefore it is crash-safe, but can be slower than the non-reliable disk buffer.

Working of the non-reliable disk-buffer is the following.
There are three levels in the queue (top to bottom):
     * qout -> this is a fast-path memory queue, we always get the messages from this
     * disk -> if there is no more space in the qout
    	   we start to move and serialize messages into the disk.
     * overflow -> if there is no more space in the disk,
                   we have to start store messages in memory,
                   turning on the flow-control if it is needed
An example configuration snippet for using non-reliable disk-buffer looks like:

       disk-buffer(
    	reliable(no)
            qout-size(64)
            disk-buf-size(1000000)
            mem-buf-size(10000)
       )

The reliable disk buffer store every message into the disk in serialized format.
but we use a watermark avoiding message dropping in case of flow control.
If the free space in the specified disk buffer become less than the mem-buf-length,
we store the messages in also in a memory queue, turning on the flow-control if it is needed.

An example configuration snippet for using reliable disk-buffer looks like:

       disk-buffer(
    	reliable(yes)
        disk-buf-size(1000000)
    	mem-buf_length(100000)
       )

Signed-off-by: Juhász Viktor <viktor.juhasz@balabit.com>