dnscache: reuse DNSCache instances in worker threads
Instead of getting rid off the per-thread DNSCache when a worker thread
exits, store them in a linked list and acquire them as a new thread starts.

The set of cached hostnames are valuable as worker threads come and
go (they exit after 10seconds of inactivity), but without this
reusing of cache instances, our DNS cache is filled again and again.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>