dbparser: fix IPv6 parser
This commit fixes the parsing of IPv4-mapped IPv6 addresses.

The r_parser_ipv6() function returned FALSE if a colon presents
after parsing the four last 8-bits as decimal values.

This is not always a good behaviour, for example:
"::ffff:200.200.200.200: 2000: Timeout, server not responding."

Instead of returning FALSE, the parser loop can be stopped
at the end of the IPv4-mapped address.

Signed-off-by: László Várady <laszlo.varady@balabit.com>