afsocket-dest: remove unused function
This commit removes the afsocket_dd_format_persist_name function,
because it violates the Clean Code paradigm.

Q: Why did I remove the function after I refactored it?
A: Before the refactor, it wasn't trivial to notice and resolve the
bool parameter, and imho reviewing that solution would be a
nightmare.

Signed-off-by: MÓZES Ádám István <adam.mozes@balabit.com>