affile: extract affile_sd_reopen_on_notify()
This is just a blind extraction of a large bulk of code that should probably
refactored to be identical to the initial opening of the file at _init()
time.

The only reason I am doing this is a bugfix, as we are not handling
NC_ERROR notifications right now, where the only sane reaction is to
try to reopen the file.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>