LogPathOptions: replace bitfields with booleans
The aim of this change is to improve the "portability" of LogPathOptions.
A bitfield is a very C like concept, it's implementation compiler
dependent.

The log_pipe_forward_msg() function takes  a LogPathOptions pointer, so
if somebody wants to call it from an other language, the LogPathOptions
type has to be redefined. If LogPathOptions contains bitfields, that
could be a problem.

I didn't find any performance penalty. We always pass around LogPathOptions
pointers, so the only dramatic change is we lose some bytes on a
stackframe.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>