native: pass super for every parse() call
We may need that pointer during a parse. The previous solution was flawed
because if the parser was cloned with clone(), the Rust parser didn't
receive the parser at all.

By passing the super pointer to each parse() call, we remove a state
and guarantee, that the pointer is always valid (basically, it's a form
of borrowing).

Signed-off-by: Tibor Benke <ihrwein@gmail.com>