dbparser: add inherit_mode to SyntheticMessage
Instead of "almost" always using an inherit_mode in case a SyntheticMessage
is used, include this as a member.

There's one use-case which ignores this attribute, when we merge in the
SyntheticMessage into an already existing LogMessage.  There we can't really
interpret this attribute.  We could perhaps create a class with inheritance
to make the model a bit more correct, however that would definitely require
a lot more boilerplate. Might make more sense in the future though.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>