cfg-args: always normalize option names
We have to normalize option names set by users.
'Normalize' means that we replace all hyphen('-') to underscore('_') in an
option name so we will find the key in the underlying hashtable and thus we
can update or read its value.

With this patch users can use both hyphen(-) and underscore (_) in
SCLs.

Fixes: #910

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>