dbparser: add create-context action
This patch adds a new possible action in the <actions> element, to create
a new correlation context out of the current message and its associated
context.

Here's an example:

<rule provider='test' id='12' class='violation'>
  <patterns>
    <pattern>simple-message-with-action-to-create-context</pattern>
  </patterns>
  <actions>
    <action trigger='match'>
      <create-context context-id='1000' context-timeout='60' context-scope='program'>
        <message inherit-properties='context'>
          <values>
            <value name='MESSAGE'>context message</value>
          </values>
        </message>
      </create-context>
    </action>
  </actions>
</rule>

Whenever our rule matches, it will create a correlation context with an
id of "1000", scope "program" and the specified content.

This can be used to "split" a state.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>