python: introduce flush-lines option
!!! EXPERIMENTAL FEATURE !!!

Syslog-ng python destination respects `flush-lines(<INT>)` option. It
uses `Py_ListObject` to queue incoming messages to the destination and
if the maximum size is reached it sends them to one of the workers as a
python list.

If `flush-lines(1)` is set, the list passed to one of the workers will
contain only one message.

If `flush-lines` is not set, python destination will use `cfg` default.

*Further improvements*:
* make python destination respect `flush-timeout` option
* do not drop messages on reload with different `flush-lines` value

*This feature was needed* in order to make developers write destinations
that can send batched events.

*Example was modified* according to the changes.

Signed-off-by: Gergő Nagy <grigori.grant@gmail.com>