ElasticSearchDestination-v2: ESMessageProcessor moved to ESClient
MessageProcessors knows how messages will be sent (bulk, single).
They need a Client object to deliver messages, so they are depend on
each other. But dealing with this dependency is not something belongs
to ElasticSearchDestination.

We have an ESClient class which can resolve the dependency and can hide
MessageProcessors from other parts. From now, we can say that
ElasticSearchDestination has an ESClient object which is responsible
for handling connection and deliver messages to ElasticSearch.

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>