cfg-lexer: avoid double free by cloning the token
cfg_token_block_get_token() returns the next token of the block or a NULL
pointer if there is no token left.

We share this token with the grammar, and the actions in the grammar
are allowed to free its parts (like the .cptr field in case of
LL_IDENTIFIER, LL_STRING, LL_BLOCK).

But we will free the content of cptr during a cfg_lexer_free_token() call.

I see two solutions:
* clone the token before giving it to the grammar (cfg_lexer_copy_token)
* don't free the individual tokens in cfg_token_block_free(), just the
array itself

As we *moved* out the tokens by cfg_token_block_get_token(), I choose the
second approach.

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>