patterndb-5.xsd: remove test_tags element
The schema had a test_tags element, which has not actually been implemented
by either syslog-ng or pdbtool. It does not make too much sense either,
while there is value in actually validating the extracted values,
tags are specified as literals, so they will only be missing because
of a patterndb bug and not much else.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>