cfg-tree: fix a potential crash in reload logic
cfg_tree_compile() is not idempotent for at least two reasons:
  * it stores the pipes to be initialized in an array, which will grow
    for each cfg_tree_compile() calls
  * it uses PIF_INLINED to indicate if a LogPipe was pasted directly into
    the config rather than cloned

But cfg_tree_start() invokes cfg_tree_compile() unconditionally, which
means that if a config is initialized multiple times, cfg_tree_compile()
will accumulate cruft in its state.

Issue #1065 on github explains one case where syslog-ng crashes on reload:
https://github.com/balabit/syslog-ng/issues/1065

This patch fixes this problem by guarding the compile function with
a variable, which ensures that the body of the function is executed only
once.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>