utf8utils: fix invalid use of UTF-8 leading byte
g_utf8_get_char_validated() converts a sequence of bytes encoded as UTF-8 to
a Unicode character (gunichar --> UTF-32), so checking for an UTF-8 leading
byte (< 0xc0) is not an option here. 0x80 can be used instead.

This is a performance optimization, where we can append the given unichar
with g_string_append_c() if it fits in 7 bits.
In all other cases, conversion is required (g_string_append_unichar()).

Signed-off-by: László Várady <laszlo.varady@balabit.com>
Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>