nvtable-serialize: don't rely on name-value pairs to come in order
In order to fix up the index in the NVTable instance, we are shadowing the
"index" (used to be called dynamic values) and update the values there.

Earlier this code simply assumed that we get the entries the order of the
index (e.g. sorted by handle value), and simply used a counter to allocate the
new index entries.

Instead of that, this patch converts this solution to just calculate the
index slot, which makes the cur_index_entry superficial.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>