logmsg: move handle fixup code to a module of its own
Instead of fixing up the handles in an NVTable in functions pretty close
to the NVTable deserialization code, do that in a clearly LogMessage
specific module.

We were depending on LogMessage concepts a lot (stuff like how an .SDATA
name-value pair looks like and the fact that there are structured
data name-value pairs where we keep the ordering), so it makes sense
to put this at a higher layer. It still pokes into NVTable internals a
lot, so clearly it is pretty tightly coupled with both LogMessage and
NVTable.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>