serialize: make typed read/write functions to be inline
Some of the serialize read/write functions are called a lot and by
hiding them in the .c file, the compiler has to embed real call
instructions, which clearly shows in the profile. With this change
I could increase LogMessage deserialization speed from 270k to 330k with
-O2.

I decided to keep the error handling parts private, those would bloat
the call-sites a lot and are rarely executed anyway (and
introduce intrusive dependencies in the header).

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>