tags-serialize: use scratch buffers in tags deserialization
Tags are serialized as strings and when reading them back, we used
serialize_read_cstring() which in turn allocates a buffer for each
string read. If a message has a couple of tags (which it does by default),
this can add to the malloc load. By using scratch buffers here we eliminate
that.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>