kvtagger: add default selector functionality
Previously, if there was no tag which matched the selector of the log message
in the database, the message was forwarded without tagging.
From now on it is possible to tag such messages. Desired name
and value pairs can be set in the external database in the following format in case of CSV database:
<default_selector>,<name1>,<value1>
<default_selector>,<name2>,<value2>
Where default_value can be specified in the configuration block of KVTagger.

Signed-off-by: Noémi Ványi <sitbackandwait@gmail.com>