kvformat/tests/test_kv_scanner: replace varargs with unhygienic macro
Note that some of the intermediate structures and helpers are removed
by the next commit.

Fix the following portability issue:
* declare and va_start() the va_list
* call a function that calls va_arg() on the same list
* try to resume the use of the va_list from the caller using va_arg()

This works on GNU/Linux, but can trigger undefined behavior according
to the man page of va_arg:

"type va_arg(va_list ap, type);
If ap is passed to a function that uses va_arg(ap,type) then the value
of ap is undefined after the return of that function.
"

Simulate varargs by passing around an array of strings.
Split 0-argument/vararg-macro as token stitching does not work here.

Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>