kvtagger: fixed reload support
When using persist config during reload, it caused a segfault when cloned parser
was used (eg.:the same parser was used in two different logpath) and the underlying
database file content invalidated before a reload operation was called.
Reference counter was also a problem.

The motivation behind falling back to the previously loaded KVTagDB is clear
(and reasonable): when the database file updated but the contens are invalid
then we have to use the previously loaded data that is available for us
(old config).

The solution is simple: we don't have to use persist_config.
During the init the database file is loaded if and only if when the
init is called during reload.
So when we have an already loaded KVTagDB instance, we don't have to
reload it (especially because if we try to reload it in the
aforementioned sitation then the init will fail for the old config
and syslog-ng will stop itself...)

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>

TODO:squash reload support

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>