configure: support java in FreeBSD
On FreeBSD there is javavm... a wrapper around java.
Every java binary is a symlink to this wrapper.
When we call javah, or javac we pass their own arguments
that are not known by the javavm (or, it calls another function
that we would expect, eg.: when we call `javac -version`,
it will call `javavm -version` and the output looks like
as if we would called `java -version` ).

Solution:
We check if the OS if FreeBSD then we won't resolve the symlink...

Fixes: #1150

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>