elastic-v2: fixed debian packaging
elastic-v2 is part of the mod-elastic package (plugin.conf is common
for elastic and elastic-v2 drivers).
Jest client is also added to this package.

This may be a little bit confusing, but I have my own reasons:
 * syslog-ng supports Elastisearch: not 1.7, 2.x, 5, but Elasticsearch
 * elastic and elastic-v2 exists just because the Elasticsearch Java clients
   are not compatible between 1.7.x and 2.x versions (and the bad thing
that the package names are almost identical, but internal representation
has been changed)
 * elastic-v2 could be packaged separately, but we have a http mode extension
   to that which is able to work with both ElasticSearch 1.x, 2.x and Elastic5

So the elastic is only for 1.7.x series, elastic-v2 could work with 2.x series
with node, transport and shield mode, and elastic-v2 http mode could work
with all versions.

References: #1118

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>