autotools: integrate Criterion unit test framework
From the documentation of Criterion:

"Criterion is a dead-simple, non-intrusive unit testing framework for C.

Most test frameworks for C require a lot of boilerplate code to set up tests
and test suites – you need to create a main, then register new test suites,
then register the tests within these suits, and finally call the right
functions.

This gives the user great control, at the unfortunate cost of simplicity.

Criterion follows the KISS principle, while keeping the control the user would
have with other frameworks.

- C99 and C++11 compatible
- Tests are automatically registered when declared
- Implements a xUnit framework structure
- A default entry point is provided, no need to declare a main unless you want
  to do special handling
- Test are isolated in their own process, crashes and signals can be reported
  and tested
- Supports parameterized tests and theories
- Progress and statistics can be followed in real time with report hooks"

Criterion is integrated as an optional "system" dependency. This can be changed
in the future; using the framework as a submodule would be a more portable
solution.
The library is searched at configure time via pkg-config, if it's not found,
Criterion related tests will be skipped.

Signed-off-by: László Várady <laszlo.varady@balabit.com>