syslog-parser: set default msg format options
The syslog parser supports msg-format related options and albeit the code
has a lot of smells in the area (beginning with all the syslog related
settings in MsgFormatOptions), we should make an effort to use them.

To do that, we need to set defaults, and initialize/destroy these options
properly.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>