syslog-parser: drop timestamp before calling the parser
The LM_TS_STAMP is already filled by the time we get to syslog-parser(),
however we do want to fully parse the date. One case that was missing was
when the input timestamp didn't have a timezone and the syslog-parser()
did have a time-zone() option specified. The correct behaviour is to
use the timezone specified in options, however since it was already parsed
earlier, it wasn't the case.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>