cfg-lexer: add cfg_lexer_format_location_tag() function
This function will format an EVTTAG, straight usable as a parameter to
one of the internal event logging API.

The name of the tag is 'location' and the content specifies the closest
file-based location to the YYLTYPE passed as argument. This can be used
to improve non-fatal error reporting during configuration parsing, as
it gives the user a much better error location.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>