Added "shield" client mode for elastic v1: This is required to be compatible with elastic v2
In case of elastic v2, if someone wants to use shield, the client mode must be "shield".
However, elastic v1 only uses "transport" for shield and gave error for "shield" mode in the config.
The different config option was confusing if someone wanted to use both elastic v1 and v2 or just switched from v1 to v2 or vice versa.

This patch fixes it, from now, if someone wants to use shield mode, he can give "shield" as client mode for both elastic v1 and v2.

Signed-off-by: Zoltan Pallagi <pzoleex@gmail.com>