kvformat: move linux-audit-parser() to a separate module
Earlier, the kv-parser.c module has contained both the kv-parser() and the
linux-audit-parser() as the code reused the same base class.

However linux-audit-parser() has additional functionality that lives in
linux-audit-scanner, which is pretty much coupled to this tiny piece that
lived in kv-parser.c.

This patch moves out the linux-audit-parser() related code to a separate
module and a subsequent patch will merge the closely coupled code into one
single module, making the code easier to follow.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>