kvformat: rename KVParsevalue to KVTransformValueFunc
The term "parse" is used for three reasons in the kvformat module:
 * kv-parser() is a syslog-ng parser, thus it has a KVParser object
 * the config grammar used by the kv-parser() module
 * then within the KVScanner, a value can be parsed into an
   alternative representation

It is too bad to have the first two already, using it a 3rd time is
more than confusing, so rename that and use the term "transform"
consistently.

E.g. we are parsing name-value pairs from the input and potentially
transform them from one representation to the next. This should
make navigation in the codebase a lot easier.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>