java-modules: remove location information from internal debug logs
Getting location information
(e.g. 'org.syslog_ng.hdfs.HdfsDestination.open:83 - Opening hdfs')
is an expensive operation.
This caused a 70% performance loss in the HDFS destination.

A "syslog-ng debug-mode" checker should be implemented on the Java side.
After this, we can revert this patch if someone finds this kind of location
information useful.

Signed-off-by: László Várady <laszlo.varady@balabit.com>