kvformat: apply testcases to both scanners
This patch attempts at enabling all testcases to both KVScannerSimple and
KVScannerGeneric in order to reveal subtle differences between the two.

The end goal is to find all the deviations, merge them and let one scanner
be the one true scanner.  At this point it is undecided which one it would
be.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>