str-repr: add the ability to specify delimiter chars
As the match_delimiter callback is in the hotpath, it makes sense to
avoid the callback overhead by checking initial character of the delimier.

This improves performance three times for a specific testcase.

This patch also groups decode related options to a struct making it easier
to extend them in the future.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>