kvformat: add support for extract-stray-words() option in kv-parser()
This patch adds support for stray words extraction, e.g. those words that
happen to be between key-value pairs and are otherwise not recognized
either as keys nor as values.

Syntax:

parser { kv-parser(prefix('.foo.') extract-stray-words-into('stray')); };

This would result in ${stray} to be set to these values. Each of the
values are properly escaped and separated by commas, so the upcoming
$(list-*) template functions would work on them.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>