python: fix segfault in Python module startup
We get remove the internal _syslogng module at each config initializations.
However if the key is not yet present, it may leave a KeyError exception
lingering in the error state, which will cause the main module initialization
to fail as well, which at the end will cause a crash.

This patch properly clears the error state, if the key does not exist.

Original patch Signed-off-by: Andras <andras.mitzki@balabit.com>

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>