logproto: limit maximum log message size (internal representation)
Previously, the meaning of the log-msg-size() option was not clearly
defined when encoding was set.

In multi-line mode, due to our conversion mechanism the buffer size
increased without limit to the maximum size.

This commit defines the meaning of the log-msg-size() option if
conversion is enabled:

If the size of incoming message is greater than log-msg-size()
(in the internal UTF-8 representation), it will be splitted.

Fixes #1318

Signed-off-by: Máté Farkas <mate.farkas@balabit.com>
Signed-off-by: László Várady <laszlo.varady@balabit.com>