tlscontext: fix segfault in X509_STORE_CTX_get_app_data(ctx)
This patch fixes a potential segfault during X.509 certificate validation.
What happens is that X509_STORE_CTX contains "application data", e.g.
the application is able to associate a series of pointers with the validation.

This uses an "id" to identify the user of the specific pointer.

This mechanism is used by the SSL library (still in openssl) to store the
pointer to the SSL session. The ID for this data is normally 0, however
if libssl.so is unloaded while libcrypto.so is not, it might happen
that this ID gets remapped to a non-zero value.

Then what leads to the crash is that libssl starts to use ID 1
to manage its SSL* pointer, while we in the validation code still
use ID 0, causing a NULL deref.

The exact reasons why this ID change happens is unclear, some apache
related information can be found here:

https://bz.apache.org/bugzilla/show_bug.cgi?id=32529

You can also find more information in github issue #1310.

Signed-off-by: Peter Gervai <grin@grin.hu>
Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>