collection-comporator: this object collects the new and deleted elements
There is an initial collection (e.g. list of the files in a directory)
There is an actual collection.
This object collects the elements existing only in the actual collection
(e.g.: newly created files), and the elements existing only in the initial
collection (e.g.: deleted files). After this comparison the actual collection
will be the initial one.

Signed-off-by: Juhász Viktor <juhasz.viktor81@gmail.com>