stats-query: reset counters after query if needed
`NamedStatsCounterItem` was introduced in `stats-query.c`, so the name
of each counter item is accessible to formatter callbacks.

It is intentionally hidden from other modules, because the scope of this PR is
not to refactor this part. Better solution would be to modify StatsCounterItem
so it stores the name of the counter, too.

Signed-off-by: Noémi Ványi <sitbackandwait@gmail.com>