affile-source: add exit_on_eof options
```
file("/dev/stdin", exit_on_eof(yes));
```

Exit on EOF option was introduced for LogTransport, so it will be easier
to include more modules in the future.

However, another layering violation was introduced by
log_proto_server_get_exit_on_eof. I selected this point for the layering
violation, because there was already one and both of the violations are
the consequence of the same root cause. So next time if someone
re-factors the proto and transport parts, it will be easier to eliminate
these violations with the same method.
Another solution would have been to store exit_on_eof in
LogProtoServerOptions and change behaviour of LogTransport based on it.
This way the value could have been accessible both from LogReader and
LogTransport.  But, it would have required more boilerplate code and
more complex "solution".

Signed-off-by: Noémi Ványi <sitbackandwait@gmail.com>