stats-query: implement indexing
Previously, the full name of a counter was unknown.
However, it is required to know it in order to support
fully querying counters using globbing.
Full name is query_key of its StatsCluster + name of its type.
E.g center.received.processed. Constructing this name is pricey,
so it is only done for each queried counter only once.

This made it possible to get rid of NamedStatsCounterItem which was
an ugly workaround and the callback have became more simple. Furthermore,
the wizardry with the query expressions are removed, too. From now on,
it is possible to call pattern_match on the filter expression specified
by the user.

The function _update_index optimizes for the lookup in the hashtable.
It makes sure that only those counters are in the index which is alive
and was queried in the past.

Signed-off-by: Noémi Ványi <sitbackandwait@gmail.com>