stats-query: initialize views
Views are introduced to aggregate existing counters.
Views are stored in a hash table. The key is the name of the view
which is same as the name of the aggregated StatsCounterItem.
The value is a ViewRecord, which contains the aggregated StatsCounterItem,
aggregator callback and the list of queries to help collect the
input counters.

Views can be used to alias counters using an empty aggregator functions.

Signed-off-by: Noémi Ványi <sitbackandwait@gmail.com>