riemann: in case of flush-lines, messages are held back infinitely
Prior to this patch syslog-ng waits until flush-lines number of
messages arrive from source before sending the bulk, without any
limit. By this patch, events are flushed in case of inactivity.
Works only for optimistic path: when riemann is connected.
This does not solve the problem when events get stuck because
server is not accessible. In this case syslog-ng will try once to send
the message(s), but will not retry until flush-lines number of events
occured.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>