file destination: invalid read
syslog-ng start a reopen_timer ivykis timer in case destination file
does not exist. The cookie of the timer is set to the destination
driver itself. When syslog-ng stops syslog-ng frees the memory behind
the cookie, but does not stop the timer. The patch explicitely calls
the cancel during deinit.
Signed-off-by: Antal Nemes <antal.nemes@balabit.com>