affile: FileReaderOptions init/deinit should encapsulate all embedded Options
FileReaderOptions incrporates a number of Options structure, yet it
was missing the init/deinit/defaults interface, so all callers
were explicitly poking into the struct, intializing members. This
is very error prone and this patch introduces the
file_reader_options_init/deinit functions, so that everything
underneath is hidden.

It removes code duplication between affile-source and wildcard-source
and makes the code a lot easier to follow.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>