gprocess: fix oracle-8-jvm related issue
The java8 oracle implementation puts two additional variables to
the environ(it is reallocated by libc). It caused crash in java
destination during our testing process.

The fix: track how many items we put to the environ array,
and free only those items. This works as libc use memcpy when realloc
the environ array and the new items appeded to the end of the
reallocated array.

Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>