Mainloop: crash when reload held back
It is possible some worker threads hold back mainloop to execute
reload for a longer period of time. In case users send stop in the
meantime, syslog-ng crashes in an assert prior this patch, because
implementation expected that mainloop had only one async call
scheduled at a time.
By this patch, syslog-ng does not crash but collects such async calls,
and execute them all when all worker threads stopped. It is also
possible users send a new reload request before reload is finished. In
this case we report back a reload is ongoing, and ignore the
request. Users might retry the reload request in a later time.
Signed-off-by: Antal Nemes <antal.nemes@balabit.com>