afsocket: add ecdh-curve-list() option to grammar
ecdh-curve-list() can be used to restrict supported curves.

The parameter is a colon separated list of curve NIDs or names,
for example 'P-521:P-384:P-256'.

For a list of available curves, execute the 'openssl ecparam -list_curves'
command.

Note that the curve list can be used only with newer versions of OpenSSL
(>= 1.0.2).

Signed-off-by: László Várady <laszlo.varady@balabit.com>