affile: add LogProtoFileReader
In a similar spirit to LogProtoFileWriter, we introduce a LogProtoFileWriter,
which is basically a set of options and a constructor that
creates the appropriate LogProtoServer instance based on these options.

Somewhat unfortunately FileReader and LogProtoFileReader are pretty similar
names, which could be fixed by changing the FileReader name to something
like FileReaderPipe, or AFFileSourceReader (akin to AFFileDestWriter).

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>