cfg: Fix double free/double close crashes
Fix the double free of level->yy_buff and double close of
level->file.include_file that can occur during a config reload if a config
file is delted.

This was observed on a system with frequent dynamic config change. The issue
occurs when a config file is deleted between the glob expansion/directory
scanning in cfg_lexer_include_file_*() functions and files being opened in
cfg_lexer_start_next_include.

Fixes #1720

Signed-off-by: Sam Stephenson <sam.stephenson@alliedtelesis.co.nz>