statistics: msg_size issue in unit testcase
The goal of the referring assertion is to have some estimate of the
memory usage of a message, and to track its change in future
developments. As it is hard to give precise memory usage, only a
reasonable lower and upper bound was given. The problem was that only
64bit systems were taken into consideration. On 32bit systems, message
size can be lower. This patch extends the boundaries so that assertion
would not be triggered on 32bit systems either.
Signed-off-by: Antal Nemes <antal.nemes@balabit.com>