cfg-lexer: allow generators to be used in generated outputs (e.g. blocks)
Block (e.g.  config generators) expansion in configuration is implemented in
the lexer, e.g.  when a token is considered, we automatically check whether
that is a generator name, and if it is, the block is generated and its
output is "included" in the current input stream.

So far, the output of a generator was not subject to this expansion, and
this patch changes this constraint.

This means that a generator can reference other generators: the system()
source can contain instances for app-parser(), both of which are generators.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>