confgen: enable passing arguments to confgen external program as env variables
Options can be set when invoking confgen external program. These options are
passed as environment variables. Note that env variables are overwritten and
cleared after the external execution.
Also note that in option names dash(-) is converted to underscore (_).

This comes handy as this way configuration can be generated not only dynamically,
but also driven from the configuration.

@module confgen context(destination) name(mygen) exec("gen.sh")

And later that can be used as:
mygen(
	option1(param1)
	option2("my other option")
)