loadbalancer: added network load balancer destination
It implements a simple round-robin load balancing between multiple destination.
Balancing is based on MSEC hashing where destinations share the same configuration
except the destination address.
The main purpose to make load balance configuration simple.

It can be used as a simple network destination and all network parameters are
recognized, the destination servers must be configured as a space separated list
of host using the "targets" option:

destination mydest {
	network-load-balancer(
		targets(server1 server2 server2) # mandatory option

		# additional network options can be set as well:
		transport(udp)
		port(3456)
	)
};

NOTE: as failover-server is not supported in this version it is left out. Otherwise
nodes are used for fail-over servers to maintain availability.

The idea and original configuration comes from Ferenc Sipos.
Big kudos for the inspiration and for his persistence on pushing me to add
this function on our flight from New York to Memphis, TN! Thx!