riemann: microseconds support
From now the event_time() option in Riemann destination configuration
can take an extra optional parameter that determines the time format
is in seconds (unixtime) or microseconds.
For example:
  event_time("$(* $UNIXTIME 1000000)" microseconds)
Dependencies:
  riemann-c-client: 1.10.0 or newer
  Riemann: 2.13 or newer
With older riemann-c-client, the microseconds option will not be
available.
Older Riemann cannot understand microseconds, though she will not
signal an error either. In that case the :time of the event will be
set to the timestamp when the message arrived to Riemann.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>